replace old Azure ServiceBus plugin with the new one.
