+++ FluentStorage.Azure.ServiceBus/Converter.cs
@@ -1,34 +1,34 @@
-﻿using Microsoft.Azure.ServiceBus;
-using System;
+﻿using System;
 using System.Collections.Generic;
+using Azure.Messaging.ServiceBus;
 using QueueMessage = FluentStorage.Messaging.QueueMessage;
 
 namespace FluentStorage.Azure.ServiceBus {
 	static class Converter {
-		public static Message ToMessage(QueueMessage message) {
+		public static ServiceBusMessage ToMessage(QueueMessage message) {
 			if (message == null)
 				throw new ArgumentNullException(nameof(message));
 
-			var result = new Message(message.Content);
+			var result = new ServiceBusMessage(message.Content);
 			if (message.Id != null) {
 				result.MessageId = message.Id;
 			}
 			if (message.Properties != null && message.Properties.Count > 0) {
 				foreach (KeyValuePair<string, string> prop in message.Properties) {
-					result.UserProperties.Add(prop.Key, prop.Value);
+					result.ApplicationProperties.Add(prop.Key, prop.Value);
 				}
 			}
 			return result;
 		}
 
-		public static QueueMessage ToQueueMessage(Message message) {
-			string id = message.MessageId ?? message.SystemProperties.SequenceNumber.ToString();
+		public static QueueMessage ToQueueMessage(ServiceBusReceivedMessage message) {
+			string id = message.MessageId;
 
-			var result = new QueueMessage(id, message.Body);
-			result.DequeueCount = message.SystemProperties.DeliveryCount;
-			if (message.UserProperties != null && message.UserProperties.Count > 0) {
-				foreach (KeyValuePair<string, object> pair in message.UserProperties) {
-					result.Properties[pair.Key] = pair.Value == null ? null : pair.Value.ToString();
+			var result = new QueueMessage(id, message.Body.ToArray());
+			result.DequeueCount = message.DeliveryCount;
+			if (message.ApplicationProperties != null && message.ApplicationProperties.Count > 0) {
+				foreach (KeyValuePair<string, object> pair in message.ApplicationProperties) {
+					result.Properties[pair.Key] = pair.Value?.ToString();
 				}
 			}
