+++ FluentStorage.Azure.ServiceBus/FluentStorage.Azure.ServiceBus.csproj
@@ -26,10 +26,9 @@
    <PropertyGroup Condition="'$(Configuration)' == 'Release'">
       <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
    </PropertyGroup>
-   <ItemGroup>
-      <PackageReference Include="Microsoft.Azure.Management.ServiceBus" Version="2.1.0" />
-      <PackageReference Include="Microsoft.Azure.ServiceBus" Version="4.1.1" />
-   </ItemGroup>
+    <ItemGroup>
+        <PackageReference Include="Azure.Messaging.ServiceBus" Version="7.16.2" />
+    </ItemGroup>
    <ItemGroup>
       <ProjectReference Include="..\FluentStorage\FluentStorage.csproj" />
    </ItemGroup>
