--- FluentStorage.Azure.ServiceBus/IAzureServiceBusMessenger.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-using FluentStorage.Messaging;
-
-namespace FluentStorage.Azure.ServiceBus {
-	/// <summary>
-	/// Azure Service Bus specific information
-	/// </summary>
-	public interface IAzureServiceBusMessenger : IMessenger {
-		/// <summary>
-		/// 
-		/// </summary>
-		/// <param name="name"></param>
-		/// <returns></returns>
-		Task CreateQueueAsync(string name);
-	}
-}
