+++ FluentStorage.sln
@@ -39,8 +39,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FluentStorage.SFTP", "Fluen
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FluentStorage.Tests.FTP", "FluentStorage.Tests.FTP\FluentStorage.Tests.FTP.csproj", "{8EC92456-416E-40F0-8126-0D2F89ABD497}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FluentStorage.Azure.Messaging.ServiceBus", "FluentStorage.Azure.Messaging.ServiceBus\FluentStorage.Azure.Messaging.ServiceBus.csproj", "{EC869154-2315-41C4-BDD2-346AE85AC7AC}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -119,10 +117,6 @@ Global
 		{8EC92456-416E-40F0-8126-0D2F89ABD497}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{8EC92456-416E-40F0-8126-0D2F89ABD497}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{8EC92456-416E-40F0-8126-0D2F89ABD497}.Release|Any CPU.Build.0 = Release|Any CPU
-		{EC869154-2315-41C4-BDD2-346AE85AC7AC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{EC869154-2315-41C4-BDD2-346AE85AC7AC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{EC869154-2315-41C4-BDD2-346AE85AC7AC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{EC869154-2315-41C4-BDD2-346AE85AC7AC}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
